requirements.txt
Django
django-allauth
django-cors-headers
django-countries
django-filter
django-heroku
django-measurement
django-nested-admin
django-phonenumber-field
django-phonenumbers
django-redis
django-rest-auth
django-versatileimagefield
djangorestframework
coreapi
drf-yasg
gunicorn
measurement
phonenumbers
Pillow
psycopg2
python-decouple
redis
Faker
whitenoise
channels_redis
channels
jet-django
pycryptodome
django-3-jet
daphne
celery
django-celery-results
django-cloudinary-storage
xhtml2pdf
python-magic
libmagic
pandas
sentry-sdk